+++ mldsa/src/native/x86_64/src/rej_uniform_eta2_avx2.c
@@ -35,7 +35,7 @@
  *            rej_eta_avx and supports multiple values for ETA via preprocessor
  *            conditionals. We move the conditionals to the frontend.
  */
-unsigned int mld_rej_uniform_eta2_avx2(
+int32_t mld_rej_uniform_eta2_avx2(
     int32_t *MLD_RESTRICT r,
     const uint8_t buf[MLD_AVX2_REJ_UNIFORM_ETA2_BUFLEN])
 {
@@ -139,7 +139,7 @@ unsigned int mld_rej_uniform_eta2_avx2(
     }
   }
 
-  return ctr;
+  return (int32_t)ctr;
 }
 
 #else /* MLD_ARITH_BACKEND_X86_64_DEFAULT && !MLD_CONFIG_MULTILEVEL_NO_SHARED \
